requirements.txt
##requirements
dash
pandas
plotly
google-api-python-client==2.112.0
uritemplate
python-dotenv